code work done for week 12

+++ week12-databases/ui.py
@@ -12,6 +12,8 @@ def display_menu():
     print("COMMAND MENU")
     print("cat  - View movies by category")
     print("year - View movies by year")
+    # STEP 7 modify display menu to provide for min command
+    print("min  - View movies by minutes")
     print("add  - Add a movie")
     print("del  - Delete a movie")
     print("exit - Exit program")
@@ -68,8 +70,31 @@ def add_movie():
 
 def delete_movie():
     movie_id = int(input("Movie ID: "))
-    db.delete_movie(movie_id)
-    print("Movie ID " + str(movie_id) + " was deleted from database.\n")
+    # Get a movie by the ID
+    movies = db.get_movie_by_id(movie_id)
+    # display to the user
+    display_movies(movies, f'Movie ID {movie_id}')
+    # ask whether user is sure they want to delete the movie
+    print("Are you sure you want to delete the above movie?")
+    # Should only delete when the user enters "y"
+    answer = input("Type 'y' to confirm: ")
+    if answer.lower() == "y":
+        db.delete_movie(movie_id)
+        print("Movie ID " + str(movie_id) + " was deleted from database.\n")
+    else:
+        # just in case they don't type y
+        print("No movie was deleted")
+
+# Step 6: add a display_movies_by_minutes() function that prompts the user
+# to enter the maximum numbe rof minutes and displays all selected movies.
+# sort by minutes in ascending order
+def display_movies_by_minutes():
+    # get minute input from user
+    minutes = int(input("Enter maximum number of minutes: "))
+    # run the function to get the matching rows
+    movies = db.get_movies_by_minutes(minutes)
+    # display to user
+    display_movies(movies, f'Runtime less than {minutes}')
         
 def main():
     db.connect()
@@ -85,6 +110,9 @@ def main():
             add_movie()
         elif command == "del":
             delete_movie()
+        # Update main() to provide for min command
+        elif command == "min":
+            display_movies_by_minutes()
         elif command == "exit":
             break
         else:
